requirements.txt
junos-eznc == 2.1.7
CherryPy == 11.0.0
ws4py == 0.4.2
watchdog == 0.8.3
isc-dhcp-leases == 0.8.1
requests == 2.18.4
PyYAML == 3.12
pika == 0.11.0
pynipap == 0.29.5
ruamel.yaml == 0.14.12
jsonpickle == 0.9.5
ansible == 2.3.2.0
peewee == 2.10.1
routes == 2.4.1
napalm == 2.2.0
napalm_base == 1.0.0
pysqlite == 2.8.3
cerberus == 1.1
